ASB-32468 fix errors related to PR 66
